# docs/architecture/api-spec.yaml
openapi: 3.1.0
info:
  title: Enginex AI API
  version: 1.0.0
  description: Auto-generated from the FastAPI app (see app/scripts/export_openapi.py) — this is the same
    schema served live at /openapi.json. WebSocket endpoints (/ws/files/{file_id}) aren't representable
    in OpenAPI 3.0 and are documented separately in docs/architecture/complete-architecture.md and services/backend/app/websockets/manager.py.
paths:
...
  /api/v1/components:
    get:
      tags:
      - Components & Libraries
      summary: List Components
      operationId: list_components_api_v1_components_get
      security:
      - HTTPBearer: []
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/app__api__v1__components__schemas__ComponentOut'
                title: Response List Components Api V1 Components Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/components/{component_id}:
    get:
      tags:
      - Components & Libraries
      summary: Get Component
      operationId: get_component_api_v1_components__component_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: component_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Component Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__api__v1__components__schemas__ComponentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/pcb/boards:
    post:
      tags:
      - PCB
      summary: Create Board
      operationId: create_board_api_v1_pcb_boards_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/pcb/boards/{board_id}:
    get:
      tags:
      - PCB
      summary: Get Board
      operationId: get_board_api_v1_pcb_boards__board_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - PCB
      summary: Update Board
      operationId: update_board_api_v1_pcb_boards__board_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBoardRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pcb/components:
    post:
      tags:
      - PCB
      summary: Create Component
      operationId: create_component_api_v1_pcb_components_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComponentRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__api__v1__pcb__schemas__ComponentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/pcb/components/{component_id}:
    get:
      tags:
      - PCB
      summary: Get Component
      operationId: get_component_api_v1_pcb_components__component_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: component_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Component Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__api__v1__pcb__schemas__ComponentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - PCB
      summary: Update Component
      operationId: update_component_api_v1_pcb_components__component_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: component_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Component Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateComponentRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__api__v1__pcb__schemas__ComponentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - PCB
      summary: Delete Component
      operationId: delete_component_api_v1_pcb_components__component_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: component_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Component Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pcb/boards/{board_id}/drc:
    post:
      tags:
      - PCB
      summary: Run Drc
      operationId: run_drc_api_v1_pcb_boards__board_id__drc_post
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pcb/boards/{board_id}/erc:
    post:
      tags:
      - PCB
      summary: Run Erc
      operationId: run_erc_api_v1_pcb_boards__board_id__erc_post
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pcb/export/gerber/{board_id}:
    get:
      tags:
      - PCB
      summary: Export Gerber
      operationId: export_gerber_api_v1_pcb_export_gerber__board_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pcb/export/bom/{board_id}:
    get:
      tags:
      - PCB
      summary: Export Bom
      operationId: export_bom_api_v1_pcb_export_bom__board_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pcb/libraries/symbols:
    get:
      tags:
      - PCB
      summary: List Library Symbols
      operationId: list_library_symbols_api_v1_pcb_libraries_symbols_get
      security:
      - HTTPBearer: []
      parameters:
      - name: library
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Library
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SymbolOut'
                title: Response List Library Symbols Api V1 Pcb Libraries Symbols Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - PCB
      summary: Create Library Symbol
      operationId: create_library_symbol_api_v1_pcb_libraries_symbols_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSymbolRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pcb/libraries/footprints:
    get:
      tags:
      - PCB
      summary: List Library Footprints
      operationId: list_library_footprints_api_v1_pcb_libraries_footprints_get
      security:
      - HTTPBearer: []
      parameters:
      - name: package_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Package Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FootprintOut'
                title: Response List Library Footprints Api V1 Pcb Libraries Footprints Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - PCB
      summary: Create Library Footprint
      operationId: create_library_footprint_api_v1_pcb_libraries_footprints_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFootprintRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FootprintOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BoardOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        file_id:
          type: string
          format: uuid
          title: File Id
        name:
          type: string
          title: Name
        width_mm:
          type: number
          title: Width Mm
        height_mm:
          type: number
          title: Height Mm
        layers_count:
          type: integer
          title: Layers Count
        data:
          additionalProperties: true
          type: object
          title: Data
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - file_id
      - name
      - width_mm
      - height_mm
      - layers_count
      - data
      - created_at
      - updated_at
      title: BoardOut
...
    CreateBoardRequest:
      properties:
        file_id:
          type: string
          format: uuid
          title: File Id
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        width_mm:
          type: number
          title: Width Mm
          default: 0
        height_mm:
          type: number
          title: Height Mm
          default: 0
        layers_count:
          type: integer
          minimum: 1.0
          title: Layers Count
          default: 2
      type: object
      required:
      - file_id
      - name
      title: CreateBoardRequest
...
    CreateComponentRequest:
      properties:
        board_id:
          type: string
          format: uuid
          title: Board Id
        reference_designator:
          type: string
          maxLength: 50
          minLength: 1
          title: Reference Designator
        footprint_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Footprint Id
        library_entry_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Library Entry Id
        position_x:
          type: number
          title: Position X
          default: 0
        position_y:
          type: number
          title: Position Y
          default: 0
        rotation_degrees:
          type: number
          title: Rotation Degrees
          default: 0
        data:
          additionalProperties: true
          type: object
          title: Data
      type: object
      required:
      - board_id
      - reference_designator
      title: CreateComponentRequest
...
    CreateFootprintRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        package_type:
          type: string
          maxLength: 100
          minLength: 1
          title: Package Type
        pads:
          items: {}
          type: array
          title: Pads
        courtyard:
          items: {}
          type: array
          title: Courtyard
        silkscreen:
          items: {}
          type: array
          title: Silkscreen
      type: object
      required:
      - name
      - package_type
      title: CreateFootprintRequest
...
    CreateSymbolRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        library:
          type: string
          maxLength: 100
          title: Library
          default: Enginex Standard
        svg_data:
          type: string
          title: Svg Data
          default: ''
        pins:
          items: {}
          type: array
          title: Pins
        meta:
          additionalProperties: true
          type: object
          title: Meta
      type: object
      required:
      - name
      title: CreateSymbolRequest
...
    FootprintOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        package_type:
          type: string
          title: Package Type
        pads:
          items: {}
          type: array
          title: Pads
        courtyard:
          items: {}
          type: array
          title: Courtyard
        silkscreen:
          items: {}
          type: array
          title: Silkscreen
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - package_type
      - pads
      - courtyard
      - silkscreen
      - created_at
      title: FootprintOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SymbolOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        library:
          type: string
          title: Library
        svg_data:
          type: string
          title: Svg Data
        pins:
          items: {}
          type: array
          title: Pins
        meta:
          additionalProperties: true
          type: object
          title: Meta
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - library
      - svg_data
      - pins
      - meta
      - created_at
      title: SymbolOut
...
    UpdateBoardRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        width_mm:
          anyOf:
          - type: number
          - type: 'null'
          title: Width Mm
        height_mm:
          anyOf:
          - type: number
          - type: 'null'
          title: Height Mm
        layers_count:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Layers Count
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
      type: object
      title: UpdateBoardRequest
    UpdateComponentRequest:
      properties:
        reference_designator:
          anyOf:
          - type: string
            maxLength: 50
            minLength: 1
          - type: 'null'
          title: Reference Designator
        position_x:
          anyOf:
          - type: number
          - type: 'null'
          title: Position X
        position_y:
          anyOf:
          - type: number
          - type: 'null'
          title: Position Y
        rotation_degrees:
          anyOf:
          - type: number
          - type: 'null'
          title: Rotation Degrees
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
      type: object
      title: UpdateComponentRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    app__api__v1__components__schemas__ComponentOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        category:
          type: string
          title: Category
        manufacturer:
          anyOf:
          - type: string
          - type: 'null'
          title: Manufacturer
        part_number:
          type: string
          title: Part Number
        datasheet_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Datasheet Url
        symbol_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Symbol Id
        footprint_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Footprint Id
        meta:
          additionalProperties: true
          type: object
          title: Meta
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - category
      - manufacturer
      - part_number
      - datasheet_url
      - symbol_id
      - footprint_id
      - meta
      - created_at
      title: ComponentOut
    app__api__v1__pcb__schemas__ComponentOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        board_id:
          type: string
          format: uuid
          title: Board Id
        reference_designator:
          type: string
          title: Reference Designator
        footprint_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Footprint Id
        library_entry_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Library Entry Id
        position_x:
          type: number
          title: Position X
        position_y:
          type: number
          title: Position Y
        rotation_degrees:
          type: number
          title: Rotation Degrees
        data:
          additionalProperties: true
          type: object
          title: Data
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - board_id
      - reference_designator
      - footprint_id
      - library_entry_id
      - position_x
      - position_y
      - rotation_degrees
      - data
      - created_at
      - updated_at
      title: ComponentOut
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.enginex.ai
  description: Production (placeholder)